Dropping rectangles
In here, I have added another sketch file to test the moving rectangles using the same techniques for the week 8 quiz because it didn't work together with the other code. I have implemented a code suggested by ChatGPT (details in the code), I then have made it into class for me to work on it.